+++ src/commands/retrieve_submission_summaries.rs
@@ -60,7 +60,7 @@ pub(crate) fn run(
         .unwrap_or_else(|| crate::project::locate_project(&cwd))?;
     let metadata = crate::project::cargo_metadata(&manifest_path, cwd)?;
     let member = metadata.query_for_member(package.as_deref())?;
-    let package_metadata = member.read_package_metadata()?;
+    let package_metadata = member.read_package_metadata(shell)?;
     crate::config::load_from_rel_path(&member.manifest_path, &package_metadata.config, shell)?;
 
     let mut atcoder_targets = indexset!();